Turn off shell vi mode
I rarely use it, and I frequently accidentally enter into it,
particularly when running tests from my editor and then hitting escape
not realizing my cursor is in the terminal. In VS Code, my test runner
doesn't handle this correctly and fails to run the tests the next time
if the shell is in normal mode.